Resolve circular dependency with lazyMobXTableSort
