+++ src/shared/components/annotation/OncoKB.spec.tsx
@@ -1,5 +1,4 @@
 import {initQueryIndicator} from "test/OncoKbMockUtils";
-import {lazyMobXTableSort} from "shared/components/lazyMobXTable/LazyMobXTable";
 import {IndicatorQueryResp} from "shared/api/generated/OncoKbAPI";
 import {IOncoKbProps} from "./OncoKB";
 import OncoKB from './OncoKB';
@@ -7,6 +6,7 @@ import React from 'react';
 import { assert } from 'chai';
 import {shallow, mount, ReactWrapper} from 'enzyme';
 import sinon from 'sinon';
+import {lazyMobXTableSort} from "../lazyMobXTable/utils";
 
 describe('OncoKB', () => {
     const props = {
